Outer Worlds Spacers Choice Edition UEVR 6DoF
This is a 6dof profile with corrected aiming. user_script.txt is configured for native rendering to eliminate some of the negative effects the best possible. The remapcontrols.dll is still used to swap the LT + Rb so you can squeeze to aim and get some haptic on shots. If you dislike these effects, you can delete that dll out of the plugins folder.